Move Bounding Box to own module
