+++ src/rtree.rs
@@ -1,3 +1,4 @@
+mod bounding_box;
 mod dimension_type;
 mod extent;
 mod rtree;
